Rename dir heroku to model-api
