--- heroku/Pipfile
@@ -1,17 +1,0 @@
-[[source]]
-name = "pypi"
-url = "https://pypi.org/simple"
-verify_ssl = true
-
-[dev-packages]
-
-[packages]
-numpy = "*"
-flask = "*"
-flask-cors = "*"
-gunicorn = "*"
-sklearn = "*"
-category-encoders = "*"
-
-[requires]
-python_version = "3.7"
